# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: MiLyfe Brain API
  description: |
    AI Agent Swarm Orchestration Platform API.
    Local-first, open-source platform for autonomous task execution
    via specialized AI agent collaboration.
  version: 2.0.0
  contact:
    name: MiLyfe Brain Team
    url: https://github.com/RealMiLyfe/MiLyfe-Brain
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8200
    description: Local development
  - url: https://staging.milyfe.ai
    description: Staging environment
  - url: https://api.milyfe.ai
    description: Production
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional API key for network-exposed deployments
  schemas:
    AgentRole:
      type: string
      enum:
        - orchestrator
        - researcher
        - coder
        - executor
        - critic
        - designer
        - writer
        - debugger
        - planner
...
    AgentState:
      type: object
      properties:
        id:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/AgentRole'
        name:
          type: string
        status:
          type: string
          enum: [idle, thinking, acting, completed, failed]
        current_task:
          type: string
          nullable: true
        thoughts:
          type: array
          items:
            type: string
        actions_taken:
          type: integer
        progress:
          type: number
          minimum: 0
          maximum: 100
        model:
          type: string
        avatar_color:
          type: string
...
    ChatMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        session_id:
          type: string
        role:
          type: string
          enum: [user, assistant, system, tool]
        content:
          type: string
        model:
          type: string
          nullable: true
        tokens_used:
          type: integer
        tool_calls:
          type: array
          items:
            type: object
        created_at:
          type: string
          format: date-time
    ChatSendRequest:
      type: object
      properties:
        message:
          type: string
          maxLength: 50000
        session_id:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        context_files:
          type: array
          items:
            type: string
      required: [message]
...
paths:
...
  /api/agents/spawn:
    post:
      tags: [Agents]
      summary: Spawn a new agent
      operationId: spawnAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  $ref: '#/components/schemas/AgentRole'
                task:
                  type: string
                model:
                  type: string
                  nullable: true
              required: [role, task]
      responses:
        '201':
          description: Agent spawned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentState'
  /api/agents/{id}:
    delete:
      tags: [Agents]
      summary: Retire an agent
      operationId: retireAgent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Agent retired
  /api/agents/{id}/message:
    post:
      tags: [Agents]
      summary: Send message to agent
      operationId: messageAgent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
              required: [message]
      responses:
        '200':
          description: Agent response
  /api/chat/send:
    post:
      tags: [Chat]
      summary: Send message with tool execution
      operationId: chatSend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatSendRequest'
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessage'
  /api/chat/history/{session_id}:
    get:
      tags: [Chat]
      summary: Get chat history
      operationId: getChatHistory
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatMessage'
...
  /api/chat/sessions/{id}:
    delete:
      tags: [Chat]
      summary: Delete a chat session
      operationId: deleteChatSession
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Session deleted
...